+++ packages/plugins/Attributions/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Feature: Use the configured `layoutTag` to identify the direction the attributions content should open in `renderType` `independent`.
 - Feature: Add configuration parameter `buttonComponent` to use a different component as the button to open the attributions if `renderType` is set to `'independent'`.
+- Feature: Expose `AttributionButton.vue` as a new export in order to use it in custom implementations.
 
 ## 1.4.0